+++ src/OrderLinesList/OrderLinesFilters.js
@@ -8,6 +8,7 @@ import {
   AcqCheckboxFilter,
   AcqDateRangeFilter,
   AcqTagsFilter,
+  AcqUnitFilter,
   BooleanFilter,
   LocationFilterContainer,
   PluggableOrganizationFilter,
@@ -244,6 +245,14 @@ function OrderLinesFilters({ activeFilters, applyFilters, disabled, materialType
         labelId="ui-orders.physical.receiptDue"
         name={FILTERS.RECEIPT_DUE}
         onChange={onChange}
+      />
+      <AcqUnitFilter
+        id={FILTERS.ACQUISITIONS_UNIT}
+        activeFilters={activeFilters[FILTERS.ACQUISITIONS_UNIT]}
+        labelId="ui-orders.order.acquisitionsUnit"
+        name={FILTERS.ACQUISITIONS_UNIT}
+        onChange={onChange}
+        disabled={disabled}
       />
       {/* <BooleanFilter
         activeFilters={activeFilters[FILTERS.CLAIM]}